requirements.txt
aiohttp
duckduckgo_search
imagehash
numpy
pillow
pydantic
PyYAML
pytest
tqdm
